MAJOR REFACTORING. Still wip.

+++ BlazorElectronics/Server/Dtos/Categories/CategoryUrlMap.cs
@@ -1,35 +1,34 @@
-using BlazorElectronics.Shared.Mutual;
+using BlazorElectronics.Shared.Categories;
 
 namespace BlazorElectronics.Server.Dtos.Categories;
 
 public sealed class CategoryUrlMap : LocallyCachedObject
 {
-    IReadOnlyDictionary<string, short> PrimaryUrlMap { get; }
-    IReadOnlyDictionary<string, IReadOnlyDictionary<short, short>> SecondaryUrlMap { get; }
-    IReadOnlyDictionary<string, IReadOnlyDictionary<short, short>> TertiaryUrlMap { get; }
+    IReadOnlyDictionary<string, int> PrimaryUrlMap { get; }
+    IReadOnlyDictionary<string, IReadOnlyDictionary<int, int>> SecondaryUrlMap { get; }
+    IReadOnlyDictionary<string, IReadOnlyDictionary<int, int>> TertiaryUrlMap { get; }
 
-    public CategoryUrlMap( Dictionary<string, short> primary, Dictionary<string, Dictionary<short, short>> secondary, Dictionary<string, Dictionary<short, short>> tertiary )
+    public CategoryUrlMap( Dictionary<string, int> primary, Dictionary<string, Dictionary<int, int>> secondary, Dictionary<string, Dictionary<int, int>> tertiary )
     {
-        PrimaryUrlMap = new Dictionary<string, short>( primary );
-
-        var secondaryTempMaps = new Dictionary<string, IReadOnlyDictionary<short, short>>();
-        var tertiaryTempMaps = new Dictionary<string, IReadOnlyDictionary<short, short>>();
+        PrimaryUrlMap = new Dictionary<string, int>( primary );
+        
+        var secondaryTempMaps = new Dictionary<string, IReadOnlyDictionary<int, int>>();
+        var tertiaryTempMaps = new Dictionary<string, IReadOnlyDictionary<int, int>>();
 
         foreach ( string url in secondary.Keys )
         {
-            Dictionary<short, short> secondaryDict = secondary[ url ];
-            secondaryTempMaps.Add( url, new Dictionary<short, short>( secondaryDict ) );
+            Dictionary<int, int> secondaryDict = secondary[ url ];
+            secondaryTempMaps.Add( url, new Dictionary<int, int>( secondaryDict ) );
         }
         foreach ( string url in tertiary.Keys )
         {
-            Dictionary<short, short> tertiaryDict = tertiary[ url ];
-            tertiaryTempMaps.Add( url, new Dictionary<short, short>( tertiaryDict ) );
+            Dictionary<int, int> tertiaryDict = tertiary[ url ];
+            tertiaryTempMaps.Add( url, new Dictionary<int, int>( tertiaryDict ) );
         }
 
         SecondaryUrlMap = secondaryTempMaps;
         TertiaryUrlMap = tertiaryTempMaps;
     }
-    
     public CategoryIdMap? GetCategoryIdMapFromUrl( List<string> urlCategories )
     {
         return urlCategories.Count switch {
@@ -42,37 +41,37 @@ public sealed class CategoryUrlMap : LocallyCachedObject
 
     CategoryIdMap? TryGetPrimaryTier( string primaryUrl )
     {
-        return PrimaryUrlMap.TryGetValue( primaryUrl, out short primaryId )
+        return PrimaryUrlMap.TryGetValue( primaryUrl, out int primaryId )
             ? new CategoryIdMap( CategoryType.PRIMARY, primaryId )
             : null;
     }
     CategoryIdMap? TryGetSecondaryTier( string primaryUrl, string secondaryUrl )
     {
-        if ( !PrimaryUrlMap.TryGetValue( primaryUrl, out short primaryId ) )
+        if ( !PrimaryUrlMap.TryGetValue( primaryUrl, out int primaryId ) )
             return null;
 
-        if ( !SecondaryUrlMap.TryGetValue( secondaryUrl, out IReadOnlyDictionary<short, short>? secondaryMap ) )
+        if ( !SecondaryUrlMap.TryGetValue( secondaryUrl, out IReadOnlyDictionary<int, int>? secondaryMap ) )
             return null;
         
-        return secondaryMap.TryGetValue( primaryId, out short secondaryId )
+        return secondaryMap.TryGetValue( primaryId, out int secondaryId )
             ? new CategoryIdMap( CategoryType.SECONDARY, secondaryId )
             : null;
     }
     CategoryIdMap? TryGetTertiaryTier( string primaryUrl, string secondaryUrl, string tertiaryUrl )
     {
-        if ( !PrimaryUrlMap.TryGetValue( primaryUrl, out short primaryId ) )
+        if ( !PrimaryUrlMap.TryGetValue( primaryUrl, out int primaryId ) )
             return null;
 
-        if ( !SecondaryUrlMap.TryGetValue( secondaryUrl, out IReadOnlyDictionary<short, short>? secondaryMap ) )
+        if ( !SecondaryUrlMap.TryGetValue( secondaryUrl, out IReadOnlyDictionary<int, int>? secondaryMap ) )
             return null;
 
-        if ( !secondaryMap.TryGetValue( primaryId, out short secondaryId ) )
+        if ( !secondaryMap.TryGetValue( primaryId, out int secondaryId ) )
             return null;
 
-        if ( !TertiaryUrlMap.TryGetValue( tertiaryUrl, out IReadOnlyDictionary<short, short>? tertiaryMap ) )
+        if ( !TertiaryUrlMap.TryGetValue( tertiaryUrl, out IReadOnlyDictionary<int, int>? tertiaryMap ) )
             return null;
 
-        return tertiaryMap.TryGetValue( secondaryId, out short tertiaryId )
+        return tertiaryMap.TryGetValue( secondaryId, out int tertiaryId )
             ? new CategoryIdMap( CategoryType.TERTIARY, tertiaryId )
             : null;
     }
